feat(order): restructure

+++ processes/order/src/main/kotlin/io/miragon/zeebe/tm/order/adapter/in/rest/model/ItemDto.kt
@@ -3,5 +3,5 @@ package io.miragon.zeebe.tm.order.adapter.`in`.rest.model
 data class ItemDto(
     val id: Long,
 
-    val quantity: Long
+    val quantity: Int
 )
